backend/requirements.txt
requests
fastapi
openai
uvicorn
slowapi
python-dotenv